Gather validation tests in SleeperPropertiesValidationTest
