+++ java/configuration/src/main/java/sleeper/configuration/properties/InstanceProperties.java
@@ -57,7 +57,7 @@ public class InstanceProperties extends SleeperProperties<InstanceProperty> {
      * Validates all UserDefinedProperties
      */
     @Override
-    protected void validate() {
+    public void validate() {
         for (UserDefinedInstanceProperty sleeperProperty : UserDefinedInstanceProperty.getAll()) {
             if (!sleeperProperty.validationPredicate().test(get(sleeperProperty))) {
                 throw new IllegalArgumentException("sleeper property: " + sleeperProperty.getPropertyName() + " is invalid");
