+++ java/configuration/src/test/java/sleeper/configuration/properties/SleeperPropertiesTest.java
@@ -168,7 +168,7 @@ class SleeperPropertiesTest {
         }
 
         @Override
-        protected void validate() {
+        public void validate() {
             // do nothing
         }
